requirements.txt
requests>=2.28.0
pyyaml>=6.0
google-generativeai>=0.3.0
openai>=1.0.0